refactor: seperated into component and container folders
