+++ app/components/index.jsx
@@ -1,11 +1,14 @@
-export { default as BasicProfile } from './BasicProfile';
 export { default as Carousel } from './Carousel';
 export { default as Champion } from './Champion';
+export { default as Champions } from './Champions';
 export { default as ChampionMasteries } from './ChampionMasteries';
 export { default as Footer } from './Footer';
-export { default as Leaderboards } from './Leaderboards';
-export { default as LiveGame } from './LiveGame';
+export { default as Leagues} from './Leagues';
 export { default as Navbar } from './Navbar';
 export { default as NotFound } from './NotFound';
-export { default as Summoner } from './Summoner';
-
+export { default as Masteries } from './Masteries';
+export { default as Match } from './Match';
+export { default as ProfileSummary } from './ProfileSummary';
+export { default as Runes } from './Runes';
+export { default as SoloFive } from './SoloFive';
+export { default as Summary } from './Summary';
