+++ app/containers/index.jsx
@@ -1,0 +1,10 @@
+export { default as BasicProfile } from './BasicProfile';
+export { default as Home } from './Home';
+export { default as LiveGame } from './LiveGame';
+export { default as Leaderboards } from './Leaderboards';
+export { default as PlayerSearch } from './PlayerSearch';
+export { default as Summoner } from './Summoner';
+export { default as SummonerTabs } from './SummonerTabs';
+export { default as SummonerTabsMobile } from './SummonerTabsMobile';
+// export { default as  } from './';
+// export { default as  } from './';
