+++ app/main.jsx
@@ -2,13 +2,9 @@ import React from 'react';
 import { render } from 'react-dom';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import Footer from './components/Footer';
 import store from './store';
-import Navbar from './components/Navbar';
-import Home from './containers/Home';
-import Summoner from './components/Summoner';
-import NotFound from './components/NotFound';
-import Leaderboards from './components/Leaderboards';
+import { Navbar, NotFound, Leaderboards, Footer } from './components';
+import { Home, Summoner } from './containers';
 
 render(
   <Provider store={store}>
